refactor: refactor directory
프로젝트 내 디렉토리 정리를 위해 enum 클래스, 테스트 클래스의 디렉토리 이동

변경사항:
- Ball, Game enum class : BaseballGame 디렉토리로 이동
- InputValidationTest : Validation 디렉토리로 이동
