+++ src/main/java/baseball/StringEnum/Game.java
@@ -1,7 +1,7 @@
 package baseball.StringEnum;
 
-import baseball.domain.Validation.Ball;
-import baseball.domain.Validation.Restart;
+import baseball.domain.BaseballGame.Ball;
+import baseball.domain.BaseballGame.Restart;
 
 public enum Game {
     START {
