+++ src/test/java/baseball/domain/Computer/ComputerTest.java
@@ -1,6 +1,6 @@
 package baseball.domain.Computer;
 
-import static baseball.domain.Validation.Ball.*;
+import static baseball.domain.BaseballGame.Ball.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
